backend/requirements.txt
sqlalchemy
fastapi
uvicorn
pydantic
python-dotenv
cors
python-jose[cryptography]
# passlib[bcrypt]
# python-multipart
python-cas
python-env